add Sona tester font by Sik of Plutiedev
Sik is working on a sound driver for Sega Genesis called "Sona".
The driver's test program (analogous to Pently's NES shell) uses
this font, which is in 8x14 and 16x28 sizes.  Both are incomplete,
and 16x28 currently lacks more glyphs (for which I currently
substitute scaled 8x14 glyphs).

Font currently lacks a name, and the glyphs are subject to change
as Sik fleshes out Sona's test shell.